requirements.txt
# Core dependencies
ultralytics>=8.0.0
ultralytics-thop>=2.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Optional - for additional features
pyyaml>=6.0
matplotlib>=3.5.0
scipy>=1.7.0

# Video processing (for saving detection videos)
imageio[ffmpeg]>=2.19.0

# GUI (only if needed)
pyqt5>=5.15.0

# For potential future voice feedback functionality
gTTS>=2.2.0
playsound>=1.2.2

# Development tools (optional)
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
